+++ admin/reports.php
@@ -10,7 +10,7 @@
                 $serviceBodies = getServiceBodiesForUser();
                 sort_on_field($serviceBodies, 'name');
                 foreach ($serviceBodies as $item) {?>
-                    <option value="<?php echo $item->id ?>"><?php echo $item->name ?> (<?php echo $item->id ?>)</option>
+                    <option value="<?php echo $item->id ?>"><?php echo $item->name ?> (<?php echo $item->id ?>) / <?php echo $item->parent_name ?> (<?php echo $item->parent_id ?>)</option>
                     <?php
                 }?>
             </select>