+++ admin/users_control.php
@@ -2,7 +2,7 @@
     <div class="modal-dialog modal-dialog-centered" role="document">
         <div class="modal-content">
             <div class="modal-header">
-                <h5 class="modal-title"><?php echo 'Add User' ?></h5>
+                <h5 class="modal-title"><span class="users_modal_title"><?php echo 'Add User' ?></span></h5>
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                     <span aria-hidden="true">&times;</span>
                 </button>
@@ -41,7 +41,7 @@
                         Service Bodies Access: <select size="10" multiple class="form-control form-control-sm" name="service_bodies" id="service_bodies">
                             <?php
                             foreach ($shareable_service_bodies as $service_body) { ?>
-                                <option value="<?php echo $service_body->id; ?>"><?php echo $service_body->name; ?></option>
+                                <option value="<?php echo $service_body->id; ?>"><?php echo $service_body->name ?> (<?php echo $service_body->id ?>) / <?php echo $service_body->parent_name ?> (<?php echo $service_body->parent_id ?>)</option>
                             <?php } ?>
                         </select>
                         <?php } ?>